+++ 19H1_Insider_Samples/CppWin32App_With_Island/SampleCppApp/XamlBridge.h
@@ -2,4 +2,8 @@
 
 #include <windows.ui.xaml.hosting.desktopwindowxamlsource.h>
 
-winrt::Windows::UI::Xaml::Hosting::DesktopWindowXamlSource CreateDesktopWindowsXamlSource(HWND parentWindow);
+extern std::vector<winrt::Windows::UI::Xaml::Hosting::DesktopWindowXamlSource> xamlSources;
+
+int MainMessageLoop(HWND hMainWnd, HACCEL hAccelTable);
+
+HWND CreateDesktopWindowsXamlSource(HWND parentWindow, DWORD dwStyle, winrt::Windows::UI::Xaml::UIElement content);
